Migrate frontend to new service tracing setup.
This alters the frontend to use the new service tracing setup.
