+++ services/frontend/pkg/revaconfig/config.go
@@ -12,7 +12,6 @@ import (
 
 // FrontendConfigFromStruct will adapt an oCIS config struct into a reva mapstructure to start a reva service.
 func FrontendConfigFromStruct(cfg *config.Config) (map[string]interface{}, error) {
-
 	webURL, err := url.Parse(cfg.PublicURL)
 	if err != nil {
 		return nil, err
@@ -71,13 +70,6 @@ func FrontendConfigFromStruct(cfg *config.Config) (map[string]interface{}, error
 	}
 
 	return map[string]interface{}{
-		"core": map[string]interface{}{
-			"tracing_enabled":      cfg.Tracing.Enabled,
-			"tracing_exporter":     cfg.Tracing.Type,
-			"tracing_endpoint":     cfg.Tracing.Endpoint,
-			"tracing_collector":    cfg.Tracing.Collector,
-			"tracing_service_name": cfg.Service.Name,
-		},
 		"shared": map[string]interface{}{
 			"jwt_secret":                cfg.TokenManager.JWTSecret,
 			"gatewaysvc":                cfg.Reva.Address, // Todo or address?
